moodle-docker/uni-analytics/requirements.txt
dash==2.16.1
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.2
mysql-connector-python==8.3.0
dash-iconify